requirements.txt
streamlit==1.31.1
openai==1.3.0
python-dotenv==1.0.0
chromadb==0.4.18 
httpx==0.24.1
PyPDF2
django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16